feat: remove unused variants of shader SVG, clearer naming
